Return stable NodeIdEdgesMaxId values in AdjacencyListGraphReader.
Currently the iterator in AdjacencyListGraphReader returns a NodeIdEdgesMaxId which is mutated on the next iteration, so if the iterator is read into another collection like a sequence, the sequence will contain many references to a holder.  After this commit AdjacencyListGraphReader constructs a new NodeIdEdgesMaxId on each iteration, which may result in more garbage collection pressure, but won't result in bizzare bugs if users ever convert the iterator to a Seq.